% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{checkDictionary}
\alias{checkDictionary}
\title{Check if dictionary meets specific requirements.}
\usage{
checkDictionary(dictionary)
}
\arguments{
\item{dictionary}{A data frame intended for feeding into the template. It should be structured according to the described conditions}
}
\value{
Logical. Returns 'TRUE' if the dictionary meets requirements for processing. Returns false otherwise
}
\description{
Verifies that the input dictionary meets the following conditions
\enumerate{
\item It is a two-column dataframe
\item Column 1 is named "flag"
\item Column 1 contains flags without starting and ending wings: « »
}
}
